+++ src/main/java/com/alibou/banking/fraud/fraud.java
@@ -1,4 +1,4 @@
-package com.alibou.banking.fraude;
+package com.alibou.banking.fraud;
 
 import com.alibou.banking.transaction.transaction;
