+++ src/main/java/com/alibou/banking/fraud/fraudType.java
@@ -1,4 +1,4 @@
-package com.alibou.banking.fraude;
+package com.alibou.banking.fraud;
 
 public enum fraudType {
     CARD_THEFT,
